+++ dwtools/Minimizers.h
@@ -2,7 +2,7 @@
 #define _Minimizers_h_
 /* Minimizers.h
  *
- * Copyright (C) 1993-2011,2015-2016 David Weenink, 2015,2017 Paul Boersma
+ * Copyright (C) 1993-2017 David Weenink, 2015,2017 Paul Boersma
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -30,17 +30,17 @@
 /*********** deferred class Minimizer **********************************/
 
 Thing_define (Minimizer, Thing) {
-	long nParameters;	/* the number of parameters */
+	integer nParameters;	/* the number of parameters */
 	double *p;			/* the parameters */
 	double minimum;		/* current minimum */
 	double *history;	/* previous minima */
 	double tolerance;	/* stop criterium */
 	Daata object;		/* reference to the object that uses this Minimizer */
-	long funcCalls;		/* the number of times 'func' has been called */
-	long success;		/* indicates whether I'm done */
-	long start;			/* start iteration series */
+	integer funcCalls;		/* the number of times 'func' has been called */
+	integer success;		/* indicates whether I'm done */
+	integer start;			/* start iteration series */
 	integer maxNumOfIterations; /* the current maximum number of iterations */
-	long iteration;     /* the total number of iterations */
+	integer iteration;     /* the total number of iterations */
 	void (*afterHook) (Minimizer me, Thing boss); /* to be called after each iteration */
 	Thing afterBoss;
 	Graphics gmonitor;		/* graphics to monitor the minimization process */
@@ -54,7 +54,7 @@ Thing_define (Minimizer, Thing) {
 	virtual void v_reset () { } /* reset the minimizer */
 };
 
-void Minimizer_init (Minimizer me, long nParameters, Daata object);
+void Minimizer_init (Minimizer me, integer nParameters, Daata object);
 /*
 	Preconditions:
 		nParameters > 0;
@@ -73,7 +73,7 @@ void Minimizer_reset (Minimizer me, const double guess[]);
  *    reset (me);
  */
 
-void Minimizer_minimize (Minimizer me, long maxNumOfIterations, double tolerance, int monitor);
+void Minimizer_minimize (Minimizer me, integer maxNumOfIterations, double tolerance, int monitor);
 /* Minimizes during maximally maxNumOfIterations. The gmonitor is initialized
  * before minimization and cleared afterwards.
  * Preconditions:
@@ -85,9 +85,9 @@ void Minimizer_minimize (Minimizer me, long maxNumOfIterations, double tolerance
  *    after each iteration: iteration++
  */
 
-void Minimizer_minimizeManyTimes (Minimizer me, long numberOfTimes, long maxIterationsPerTime, double tolerance);
+void Minimizer_minimizeManyTimes (Minimizer me, integer numberOfTimes, integer maxIterationsPerTime, double tolerance);
 
-void Minimizer_drawHistory (Minimizer me, Graphics g, long itmin, long itmax, double minimum, double maximum, int garnish);
+void Minimizer_drawHistory (Minimizer me, Graphics g, integer itmin, integer itmax, double minimum, double maximum, int garnish);
 
 double Minimizer_getMinimum (Minimizer me);
 
@@ -106,7 +106,7 @@ Thing_define (LineMinimizer, Minimizer) {
 	//virtual void v_linmin (double p[], double fp, double direction[], double *fret);	 // David, is dit correct? ja
 };
 
-void LineMinimizer_init (LineMinimizer me, long nParameters, Daata object, double (*func) (Daata object, const double p[]));
+void LineMinimizer_init (LineMinimizer me, integer nParameters, Daata object, double (*func) (Daata object, const double p[]));
 
 /******************  class SteepestDescentMinimizer**************************/
 
@@ -124,14 +124,14 @@ Thing_define (SteepestDescentMinimizer, Minimizer) {
 		override;
 };
 
-autoSteepestDescentMinimizer SteepestDescentMinimizer_create (long nParameters, Daata object, double (*func) (Daata object, const double p[]), void (*dfunc) (Daata object, const double p[], double dp[]));
+autoSteepestDescentMinimizer SteepestDescentMinimizer_create (integer nParameters, Daata object, double (*func) (Daata object, const double p[]), void (*dfunc) (Daata object, const double p[], double dp[]));
 
 
 /**********  class VDSmagtMinimizer ********************************/
 
 typedef struct structVDSmagtMinimizer_parameters {
 	double lineSearchGradient;
-	long lineSearchMaxNumOfIterations;
+	integer lineSearchMaxNumOfIterations;
 } *VDSmagtMinimizer_parameters;
 
 Thing_define (VDSmagtMinimizer, Minimizer) {
@@ -139,7 +139,7 @@ Thing_define (VDSmagtMinimizer, Minimizer) {
 	void  (*dfunc) (Daata object, const double p[], double dp[]);
 	double *dp;
 	double lineSearchGradient;
-	long lineSearchMaxNumOfIterations;
+	integer lineSearchMaxNumOfIterations;
 	double gr0, gropt, df, alplim, alpha, dalpha, alphamin;
 	double *pc;	/* position of current point */
 	double *gc;	/* gradient of current point */
@@ -149,8 +149,8 @@ Thing_define (VDSmagtMinimizer, Minimizer) {
 	double *grst; /* gradient for first iteration after restart */
 	double fc, grc, fch, gr2s, temp, grs, beta, gcg0;
 	double gamma, gamma_in, f0, gsq, gopt_sq;
-	long lineSearch_iteration, flag, again, one_up, restart;
-	long restart_flag;
+	integer lineSearch_iteration, flag, again, one_up, restart;
+	integer restart_flag;
 
 	void v_destroy () noexcept
 		override;
@@ -160,6 +160,6 @@ Thing_define (VDSmagtMinimizer, Minimizer) {
 		override;
 };
 
-autoVDSmagtMinimizer VDSmagtMinimizer_create (long dimension, Daata object, double (*func) (Daata object, const double p[]), void (*dfunc) (Daata object, const double p[], double dp[]));
+autoVDSmagtMinimizer VDSmagtMinimizer_create (integer dimension, Daata object, double (*func) (Daata object, const double p[]), void (*dfunc) (Daata object, const double p[], double dp[]));
 
 #endif /* _Minimizer_h_ */
